Исправил аннтоции, добавил понятные статус коды, HTTPException. И некоторые места в коде.

+++ src/routers/auth/code/register.py
@@ -1,6 +1,5 @@
 
-
-from fastapi import APIRouter, Response, Depends, HTTPException
+from fastapi import APIRouter, Request, Response, Depends, HTTPException
 
 from src.database.bases.auth_base import regstr, check
 from src.routers.auth.schemas import Reg, Auth
@@ -11,7 +10,7 @@ from auth2 import auth
 router = APIRouter(tags=['Register'])
 
 
-async def depend_register_user(data: Reg, response: Response):
+async def depend_register_user(data: Reg, response: Response) -> dict:
      await decor.registration_name(arg=data.name)
      
      insert = await regstr.insert_new_data(data=data.model_dump())
@@ -20,13 +19,12 @@ async def depend_register_user(data: Reg, response: Response):
                username=data.name, 
                response=response
           )
-     
-     raise HTTPException(status_code=170, detail='This name already taken!')
+     raise HTTPException(status_code=440, detail='This name already taken!')
 
 
 
 
-async def depend_auth(data: Auth, response: Response):
+async def depend_auth(data: Auth, response: Response) -> dict:
      await decor.registration_name(arg=data.name)
      
      passw = await check.check_password(
@@ -34,21 +32,25 @@ async def depend_auth(data: Auth, response: Response):
           password=data.password
      )
      if passw is False:
-          return {'status': 50, 'detail': 'Invalid name or password!'}
+          raise HTTPException(status_code=442, detail='Invalid name or password!')
 
-     return await auth.get_access_refresh_tokens(
+     await auth.get_access_refresh_tokens(
           username=data.name,
           response=response
      )
+     return {'status': 206, 'detail': 'Success!'}
      
      
      
      
-async def depend_out_user(response: Response):
-     response.delete_cookie(key='access_token', httponly=True)
-     response.delete_cookie(key='refresh_token', httponly=True)
+async def depend_out_user(response: Response, request: Request) -> dict:
+     if request.cookies.get('access_token') and request.cookies.get('refresh_token'):
+          response.delete_cookie(key='access_token', httponly=True)
+          response.delete_cookie(key='refresh_token', httponly=True)
+          
+          return {'status': 207, 'access_token': None, 'refresh_token': None}
      
-     return {'status': 550, 'access_token': None, 'refresh_token': None}
+     raise HTTPException(status_code=443, detail='You dont auth.')
